\name{ISO_CountCodes}
\alias{ISO_CountCodes}
\docType{data}
\title{
ISO Country names and codes
}
\description{
A data frame of ISO Country names and codes with three variables; country name, two letter and three country letter codes.
}
\usage{data("ISO_CountCodes")}
\format{
  \itemize{
  \item countName is the full country English name
  \item iso2code is a two-letter country code
  \item iso3Code is a three-letter country code
  }
}

\source{
 \url{https://www.iso.org/obp/ui#search}
}

\examples{
data(ISO_CountCodes)
}
\keyword{datasets}
